Refactor battery handling so v5 PMIC code and pre v5 simple battery voltage divider works cleanly.
