+++ app/CMakeLists.txt
@@ -38,14 +38,17 @@ add_subdirectory(src/events)
 add_subdirectory(src/managers)
 add_subdirectory(src/drivers)
 add_subdirectory(src/ble)
-add_subdirectory(src/battery)
 add_subdirectory(src/images)
 add_subdirectory(src/images/fonts)
-add_subdirectory(src/fuel_gauge)
-
 include_directories(src/)
 include_directories(src/ui)
 include_directories(src/applications)
+
+if (CONFIG_DT_HAS_NORDIC_NPM1300_ENABLED)
+add_subdirectory(src/fuel_gauge)
+else()
+add_subdirectory(src/basic_battery)
+endif()
 
 target_sources(app PRIVATE src/main.c)
 target_sources(app PRIVATE src/zsw_clock.c)
