+++ app/boards/arm/zswatch_nrf5340/zswatch_nrf5340_cpuapp_5.conf
@@ -39,3 +39,7 @@ CONFIG_PM_PARTITION_SIZE_LITTLEFS=0x200000
 CONFIG_NRF_FUEL_GAUGE=y
 
 CONFIG_APPLICATIONS_USE_BATTERY=n
+
+# Temporary due to current flash having larger pages than NVS can handle.
+CONFIG_FCB=y
+CONFIG_SETTINGS_FCB=y
