+++ app/boards/debug.conf
@@ -24,4 +24,3 @@ CONFIG_ZSW_SENSORS_LOG_LEVEL_WRN=y
 CONFIG_ZSW_BLE_LOG_LEVEL_WRN=y
 CONFIG_ZSW_SETTINGS_APP_LOG_LEVEL_WRN=y
 CONFIG_ZSW_APP_LOG_LEVEL_DBG=y
-CONFIG_NRF_FUEL_GAUGE=y
