+++ app/boards/npm1300_ek.conf
@@ -1,0 +1,1 @@
+CONFIG_NRF_FUEL_GAUGE=y
