+++ app/boards/npm1300_ek.overlay
@@ -1,7 +1,4 @@
-/*
- * Copyright (C) 2023 Nordic Semiconductor ASA
- * SPDX-License-Identifier: Apache-2.0
- */
+// For use with nRF5350DK
 
  #include <dt-bindings/regulator/npm1300.h>
  #include <zephyr/dt-bindings/input/input-event-codes.h>
@@ -68,14 +65,5 @@
 			 thermistor-beta = <3380>;
 			 charging-enable;
 		 };
-		/*
-		 npm1300_ek_buttons: buttons {
-			 compatible = "gpio-keys";
-			 pmic_button0: pmic_button_0 {
-				 gpios = < &npm1300_ek_gpio 0 GPIO_ACTIVE_HIGH>;
-				 label = "Pmic button switch 0";
-			 zephyr,code = <INPUT_KEY_0>;
-			 };
-		 };*/
 	 };
  };
