+++ app/boards/nrf5340dk_nrf5340_cpuapp.conf
@@ -57,4 +57,3 @@ CONFIG_PM_PARTITION_SIZE_SETTINGS_STORAGE=0x100000
 # how external flash shall be utilized. nRF5340dk has QSPI flash, hence
 # similar to rev 4.
 CONFIG_ZSWATCH_PCB_REV=4
-CONFIG_NRF_FUEL_GAUGE=y
