+++ app/prj.conf
@@ -114,9 +114,8 @@ CONFIG_FLASH_PAGE_LAYOUT=y
 CONFIG_FLASH_MAP_LABELS=y
 
 CONFIG_SETTINGS=y
-#CONFIG_NVS=y
-CONFIG_FCB=y
-CONFIG_SETTINGS_FCB=y
+CONFIG_NVS=y
+CONFIG_SETTINGS_NVS=y
 
 # Bluetooth
 CONFIG_BT=y
