--- app/src/events/chg_event.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <stdbool.h>
-
-struct chg_state_event {
-    bool is_charging;
-};
