--- app/src/managers/zsw_battery_manager.h
@@ -1,25 +1,0 @@
-/*
- * This file is part of ZSWatch project <https://github.com/jakkra/ZSWatch/>.
- * Copyright (c) 2023 Jakob Krantz.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, version 3.
- *
- * This program is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-#pragma once
-
-/** @brief          Sample the battery voltage, output in mV and percent.
- *  @param mV
- *  @param percent
- *  @return         0 when successful
-*/
-int zsw_battery_manager_sample_battery(int *mV, int *percent);
